requirements.txt
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
amqp==5.0.5
apiclient==1.0.4
asgiref==3.3.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
django-celery-results==2.0.1
django-common-helpers==0.9.2
django-crispy==1.8.1
django-crispy-forms==1.11.2
django-cron==0.5.1
django-utils-six==2.0
djangorestframework==3.12.2
eventemitter==0.2.0
gevent==21.1.2
google-api-core==1.26.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
httplib2==0.19.0
idna==2.10
kombu==5.0.2
multidict==5.1.0
oauth2client==4.1.3
packaging==20.9
Pillow==8.1.0
postgres==3.0.0
prompt-toolkit==3.0.18
protobuf==3.15.6
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.10.1
pyparsing==2.4.7
python-socks==1.2.3
pytz==2020.5
redis==3.5.3
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
steam==1.2.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vdf==3.3
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
